+++ History.md
@@ -1,3 +1,9 @@
+0.2.0 / 2014-11-18
+==================
+* Renaming styles to types
+* adding type/default meta
+* breaking change; renaming `StyleGuide.add(type, filename)` to `Styleguide.load(filename)`
+
 0.1.0 / 2014-10-29
 ==================
 * initial commit